+++ src/components/scenes/CoinRankingDetailsScene.tsx
@@ -1,10 +1,10 @@
-import { useIsFocused } from '@react-navigation/native'
 import { EdgeCurrencyWallet } from 'edge-core-js'
 import * as React from 'react'
 import { ActivityIndicator, View } from 'react-native'
 import FastImage from 'react-native-fast-image'
 import Feather from 'react-native-vector-icons/Feather'
 import Ionicons from 'react-native-vector-icons/Ionicons'
+import { sprintf } from 'sprintf-js'
 
 import { createWallet, getUniqueWalletName } from '../../actions/CreateWalletActions'
 import { getFirstOpenInfo } from '../../actions/FirstOpenActions'
@@ -21,7 +21,7 @@ import { lstrings } from '../../locales/strings'
 import { getStakePlugins } from '../../plugins/stake-plugins/stakePlugins'
 import { filterStakePolicies, StakePolicy } from '../../plugins/stake-plugins/types'
 import { defaultWalletStakingState } from '../../reducers/StakingReducer'
-import { getDefaultFiat } from '../../selectors/SettingsSelectors'
+import { getCoingeckoFiat, getDefaultFiat } from '../../selectors/SettingsSelectors'
 import { asCoinRankingData, CoinRankingData, CoinRankingDataPercentChange } from '../../types/coinrankTypes'
 import { useDispatch, useSelector } from '../../types/reactRedux'
 import { EdgeAppSceneProps, NavigationBase } from '../../types/routerTypes'
@@ -32,20 +32,19 @@ import { getBestApyText, isStakingSupported } from '../../util/stakeUtils'
 import { getUkCompliantString } from '../../util/ukComplianceUtils'
 import { formatLargeNumberString as formatLargeNumber } from '../../util/utils'
 import { IconButton } from '../buttons/IconButton'
+import { AlertCardUi4 } from '../cards/AlertCard'
 import { SwipeChart } from '../charts/SwipeChart'
 import { EdgeAnim, fadeInLeft } from '../common/EdgeAnim'
 import { SceneWrapper } from '../common/SceneWrapper'
 import { Airship, showError } from '../services/AirshipInstance'
 import { cacheStyles, Theme, useTheme } from '../services/ThemeContext'
 import { EdgeText } from '../themed/EdgeText'
-import { COINGECKO_SUPPORTED_FIATS } from './CoinRankingScene'
 
 type CoinRankingDataValueType = string | number | CoinRankingDataPercentChange | undefined
 
 export interface CoinRankingDetailsParams {
-  assetId?: string
+  assetId: string
   coinRankingData?: CoinRankingData
-  fiatCurrencyCode: string
 }
 
 interface Props extends EdgeAppSceneProps<'coinRankingDetails'> {}
@@ -103,7 +102,7 @@ const CoinRankingDetailsSceneComponent = (props: Props) => {
   const styles = getStyles(theme)
   const dispatch = useDispatch()
   const { route, navigation } = props
-  const { assetId, fiatCurrencyCode, coinRankingData: initCoinRankingData } = route.params
+  const { assetId, coinRankingData: initCoinRankingData } = route.params
 
   const account = useSelector(state => state.core.account)
   const exchangeRates = useSelector(state => state.exchangeRates)
@@ -111,19 +110,19 @@ const CoinRankingDetailsSceneComponent = (props: Props) => {
 
   const currencyConfigMap = useWatch(account, 'currencyConfig')
   const currencyWallets = useWatch(account, 'currencyWallets')
-  const isFocused = useIsFocused()
 
   const [countryCode] = useAsyncValue(async () => (await getFirstOpenInfo()).countryCode)
 
   // In case the user changes their default fiat while viewing this scene, we
   // want to go back since the parent scene handles fetching data.
   const defaultFiat = useSelector(state => getDefaultFiat(state))
-  const supportedFiat = COINGECKO_SUPPORTED_FIATS[defaultFiat as keyof typeof COINGECKO_SUPPORTED_FIATS] != null ? defaultFiat : 'USD'
+  const coingeckoFiat = useSelector(state => getCoingeckoFiat(state))
+
   const [fetchedCoinRankingData] = useAsyncValue(async () => {
     if (assetId == null) {
       throw new Error('No currencyCode or coinRankingData provided')
     }
-    const response = await fetchRates(`v2/coinrankAsset/${assetId}?fiatCode=iso:${supportedFiat}`)
+    const response = await fetchRates(`v2/coinrankAsset/${assetId}?fiatCode=iso:${coingeckoFiat}`)
     if (!response.ok) {
       const text = await response.text()
       throw new Error(`Unable to fetch coin ranking data. ${text}`)
@@ -133,7 +132,7 @@ const CoinRankingDetailsSceneComponent = (props: Props) => {
     const crData = asCoinRankingData(json.data)
 
     return crData
-  }, [assetId, supportedFiat])
+  }, [assetId, coingeckoFiat])
 
   const coinRankingData = fetchedCoinRankingData ?? initCoinRankingData
 
@@ -163,8 +162,6 @@ const CoinRankingDetailsSceneComponent = (props: Props) => {
     return out
   }, [account.currencyConfig, coinRankingData, currencyCode])
 
-  const initFiat = React.useState<string>(fiatCurrencyCode)[0]
-
   /** Find all wallets that can hold this asset */
   const matchingWallets = React.useMemo(
     () => Object.values(currencyWallets).filter(wallet => edgeAssets.some(asset => asset.pluginId === wallet.currencyInfo.pluginId)),
@@ -186,15 +183,6 @@ const CoinRankingDetailsSceneComponent = (props: Props) => {
       ).length > 0
     )
   })
-
-  React.useEffect(() => {
-    if (isFocused && initFiat !== supportedFiat) {
-      // Take this stale scene off the stack
-      navigation.pop()
-      // Force a refresh & refetch
-      navigation.navigate('coinRanking')
-    }
-  }, [initFiat, isFocused, navigation, supportedFiat])
 
   React.useEffect(() => {
     // Initialize staking state
@@ -291,7 +279,7 @@ const CoinRankingDetailsSceneComponent = (props: Props) => {
       case 'marketCap':
         return `${formatFiatString({
           fiatAmount: baseString
-        })} ${supportedFiat}`
+        })} ${coingeckoFiat}`
       case 'rank':
         return `#${baseString}`
       case 'marketCapChange24h':
@@ -300,13 +288,13 @@ const CoinRankingDetailsSceneComponent = (props: Props) => {
       case 'allTimeHigh': {
         const fiatString = `${formatFiatString({
           fiatAmount: baseString
-        })} ${supportedFiat}`
+        })} ${coingeckoFiat}`
         return coinRankingData?.allTimeHighDate != null ? `${fiatString} - ${toLocaleDate(new Date(coinRankingData.allTimeHighDate))}` : fiatString
       }
       case 'allTimeLow': {
         const fiatString = `${formatFiatString({
           fiatAmount: baseString
-        })} ${supportedFiat}`
+        })} ${coingeckoFiat}`
         return coinRankingData?.allTimeLowDate != null ? `${fiatString} - ${toLocaleDate(new Date(coinRankingData.allTimeLowDate))}` : fiatString
       }
       default:
@@ -504,7 +492,7 @@ const CoinRankingDetailsSceneComponent = (props: Props) => {
             <FastImage style={styles.icon} source={imageUrlObject} />
             <EdgeText style={styles.title}>{`${currencyName} (${currencyCode})`}</EdgeText>
           </EdgeAnim>
-          <SwipeChart assetId={coinRankingData.assetId} currencyCode={currencyCode} fiatCurrencyCode={initFiat} />
+          <SwipeChart assetId={coinRankingData.assetId} />
           {edgeAssets.length <= 0 ? null : (
             <View style={styles.buttonsContainer}>
               <IconButton label={lstrings.title_buy} onPress={handleBuyPress}>
@@ -531,6 +519,13 @@ const CoinRankingDetailsSceneComponent = (props: Props) => {
                 <Ionicons name="swap-horizontal" size={theme.rem(2)} color={theme.primaryText} />
               </IconButton>
             </View>
+          )}
+          {defaultFiat === coingeckoFiat ? null : (
+            <AlertCardUi4
+              type="warning"
+              title={lstrings.coin_rank_currency_rates_warning_title}
+              body={sprintf(lstrings.coin_rank_currency_rates_warning_message_2s, coingeckoFiat, defaultFiat)}
+            />
           )}
           <View style={styles.columns}>
             <View style={styles.column}>{renderRows(coinRankingData, COLUMN_LEFT_DATA_KEYS)}</View>