+++ src/components/scenes/Loans/LoanCreateScene.tsx
@@ -14,7 +14,6 @@ import { PaymentMethod } from '../../../controllers/action-queue/PaymentMethod'
 import { useAllTokens } from '../../../hooks/useAllTokens'
 import { useAsyncEffect } from '../../../hooks/useAsyncEffect'
 import { useHandler } from '../../../hooks/useHandler'
-import { useCurrencyFiatRate } from '../../../hooks/useTokenDisplayData'
 import { useUrlHandler } from '../../../hooks/useUrlHandler'
 import { useWalletBalance } from '../../../hooks/useWalletBalance'
 import { useWalletName } from '../../../hooks/useWalletName'
@@ -206,8 +205,7 @@ export const LoanCreateScene = (props: Props) => {
     isoFiatCurrencyCode: defaultIsoFiat
   })
 
-  const isUserInputComplete =
-    srcWallet != null && (destWallet != null || destBankId != null) && !zeroString(borrowAmountFiat) && !zeroString(collateralToFiatRate)
+  const isUserInputComplete = srcWallet != null && (destWallet != null || destBankId != null) && !zeroString(borrowAmountFiat) && collateralToFiatRate !== 0
   let totalRequiredCollateralNativeAmount = !isUserInputComplete
     ? '0'
     : truncateDecimals(mul(collateralExchangeMultiplier, div(totalRequiredCollateralFiat, collateralToFiatRate, DECIMAL_PRECISION)), 0)
@@ -475,3 +473,10 @@ const getStyles = cacheStyles((theme: Theme) => ({
     textAlign: 'left'
   }
 }))
+
+const useCurrencyFiatRate = ({ currencyCode, isoFiatCurrencyCode }: { currencyCode?: string; isoFiatCurrencyCode?: string }): number => {
+  return useSelector(state => {
+    if (currencyCode == null || isoFiatCurrencyCode == null) return 0
+    else return state.exchangeRates[`${currencyCode}_${isoFiatCurrencyCode}`] ?? 0
+  })
+}